+++ arch/riscv/kernel/process.c
@@ -258,9 +258,9 @@ int copy_thread(struct task_struct *p, const struct kernel_clone_args *args)
 		/* Supervisor/Machine, irqs on: */
 		childregs->status = SR_PP | SR_PIE;
 
-		p->thread.s[0] = (unsigned long)args->fn;
-		p->thread.s[1] = (unsigned long)args->fn_arg;
-		p->thread.ra = (unsigned long)ret_from_fork_kernel_asm;
+		p->thread.s[0] = (uintptr_t)args->fn;
+		p->thread.s[1] = (uintptr_t)args->fn_arg;
+		p->thread.ra = (uintptr_t)ret_from_fork_kernel_asm;
 	} else {
 		/* allocate new shadow stack if needed. In case of CLONE_VM we have to */
 		ssp = shstk_alloc_thread_stack(p, args);
@@ -278,12 +278,12 @@ int copy_thread(struct task_struct *p, const struct kernel_clone_args *args)
 		if (clone_flags & CLONE_SETTLS)
 			childregs->tp = tls;
 		childregs->a0 = 0; /* Return value of fork() */
-		p->thread.ra = (unsigned long)ret_from_fork_user_asm;
+		p->thread.ra = (uintptr_t)ret_from_fork_user_asm;
 	}
 	p->thread.riscv_v_flags = 0;
 	if (has_vector() || has_xtheadvector())
 		riscv_v_thread_alloc(p);
-	p->thread.sp = (unsigned long)childregs; /* kernel sp */
+	p->thread.sp = (uintptr_t)childregs; /* kernel sp */
 	return 0;
 }
