+++ block/partitions/core.c
@@ -153,7 +153,7 @@ static struct parsed_partitions *check_partition(struct gendisk *hd)
 	if (res > 0) {
 		printk(KERN_INFO "%s", state->pp_buf);
 
-		free_page((unsigned long)state->pp_buf);
+		free_page((uintptr_t)state->pp_buf);
 		return state;
 	}
 	if (state->access_beyond_eod)
@@ -169,7 +169,7 @@ static struct parsed_partitions *check_partition(struct gendisk *hd)
 		printk(KERN_INFO "%s", state->pp_buf);
 	}
 
-	free_page((unsigned long)state->pp_buf);
+	free_page((uintptr_t)state->pp_buf);
 	free_partitions(state);
 	return ERR_PTR(res);
 }
