+++ crypto/aead.c
@@ -34,7 +34,7 @@ static int setkey_unaligned(struct crypto_aead *tfm, const u8 *key,
 	if (!buffer)
 		return -ENOMEM;
 
-	alignbuffer = (u8 *)ALIGN((unsigned long)buffer, alignmask + 1);
+	alignbuffer = (u8 *)ALIGN((uintptr_t)buffer, alignmask + 1);
 	memcpy(alignbuffer, key, keylen);
 	ret = crypto_aead_alg(tfm)->setkey(tfm, alignbuffer, keylen);
 	kfree_sensitive(buffer);
