+++ crypto/cipher.c
@@ -31,7 +31,7 @@ static int setkey_unaligned(struct crypto_cipher *tfm, const u8 *key,
 	if (!buffer)
 		return -ENOMEM;
 
-	alignbuffer = (u8 *)ALIGN((unsigned long)buffer, alignmask + 1);
+	alignbuffer = (u8 *)ALIGN((uintptr_t)buffer, alignmask + 1);
 	memcpy(alignbuffer, key, keylen);
 	ret = cia->cia_setkey(crypto_cipher_tfm(tfm), alignbuffer, keylen);
 	kfree_sensitive(buffer);
@@ -66,7 +66,7 @@ static inline void cipher_crypt_one(struct crypto_cipher *tfm,
 	if (unlikely(((unsigned long)dst | (unsigned long)src) & alignmask)) {
 		unsigned int bs = crypto_cipher_blocksize(tfm);
 		u8 buffer[MAX_CIPHER_BLOCKSIZE + MAX_CIPHER_ALIGNMASK];
-		u8 *tmp = (u8 *)ALIGN((unsigned long)buffer, alignmask + 1);
+		u8 *tmp = (u8 *)ALIGN((uintptr_t)buffer, alignmask + 1);
 
 		memcpy(tmp, src, bs);
 		fn(crypto_cipher_tfm(tfm), tmp, tmp);
