+++ crypto/skcipher.c
@@ -133,7 +133,7 @@ finish:
 	if (walk->buffer != walk->page)
 		kfree(walk->buffer);
 	if (walk->page)
-		free_page((unsigned long)walk->page);
+		free_page((uintptr_t)walk->page);
 
 out:
 	return res;
@@ -388,7 +388,7 @@ static int skcipher_setkey_unaligned(struct crypto_skcipher *tfm,
 	if (!buffer)
 		return -ENOMEM;
 
-	alignbuffer = (u8 *)ALIGN((unsigned long)buffer, alignmask + 1);
+	alignbuffer = (u8 *)ALIGN((uintptr_t)buffer, alignmask + 1);
 	memcpy(alignbuffer, key, keylen);
 	ret = cipher->setkey(tfm, alignbuffer, keylen);
 	kfree_sensitive(buffer);
